fixed custom autospin

+++ Assets/Scripts/APIs/SocketModel.cs
@@ -1,10 +1,6 @@
-using System.Collections;
 using System.Collections.Generic;
-using UnityEngine;
 using System;
-using Newtonsoft.Json.Linq;
-using System.Runtime.Serialization;
-using Newtonsoft.Json;
+
 
 public class SocketModel 
 {
@@ -107,7 +103,7 @@ public class Symbol
 {
     public int ID { get; set; }
     public string Name { get; set; }
-    public List<List<int>> Multiplier { get; set;}
+    public List<List<double>> Multiplier { get; set;}
     public object defaultAmount { get; set; }
     public object symbolsCount { get; set; }
     public object increaseValue { get; set; }
